Move HashManager to dedicated module, change test suite name
